refactor(frontend): move staff pages into a Staff folder
StaffTeams and StaffTrials now sit next to their layout as Staff/Teams and
Staff/Trials, following the Account folder convention where children drop the
parent prefix.
